Partitioned adapter supports two item tuples as keys (#214)
There was a problem that `get_all` when asking for a two element tuple key (ie
`{"foo", 1}) would only consider the first element of the tuple to be the key,
and therefore may ask for the data on the wrong node. This seems to have been
done to support put_(new_)all which puts the key in the first element and values in
the second.

I've passed along the action to the `group_keys_by_node` function so we're able
to know which structure to expect.