Fix race condition in multilevel cache replication
This commit fixes a race condition that is likely to happen when
one of the layers for a multilevel cache has to do some networking.

The issue arises from the calling `get` to see if a key exists, and
then calling `ttl` when replicating the key to other layers. This
means that the calls are happening in a non-atomic way, which in
turn means that by the time `ttl` gets called the key might have
expired already (the race condition).

The fix is detecting that the key has expired by checking for `nil`
in the `tll` call result, and in this case not replicating the
cached entry to the other layers.